Remove usecases for ward and trust retrieval since they've been moved over to gateways
